First steps toward mac gelpia

+++ requirements/build.sh
@@ -20,13 +20,17 @@ CRLIBM=$SCRIPT_LOCATION/crlibm/crlibm.tar.gz
 mkdir -p $SOURCE_LOCATION
 
 # Rust
-echo Installing Rust
-cd $SOURCE_LOCATION
-rm -f rust_nightly.tar.gz
-wget https://static.rust-lang.org/dist/rust-1.37.0-x86_64-unknown-linux-gnu.tar.gz -O rust_nightly.tar.gz &>> $SCRIPT_LOCATION/log.txt
-mkdir -p rust_nightly && tar -xf rust_nightly.tar.gz -C rust_nightly --strip-components 1
-cd rust_nightly
-./install.sh --prefix=$SCRIPT_LOCATION &>> $SCRIPT_LOCATION/log.txt
+if [[ "$OSTYPE" == "linux-gnu"* ]]
+then
+    echo Installing Rust
+    cd $SOURCE_LOCATION
+    rm -f rust_nightly.tar.gz
+    wget https://static.rust-lang.org/dist/rust-1.37.0-x86_64-unknown-linux-gnu.tar.gz -O rust_nightly.tar.gz >> $SCRIPT_LOCATION/log.txt 2>&1
+    mkdir -p rust_nightly && tar -xf rust_nightly.tar.gz -C rust_nightly --strip-components 1
+    cd rust_nightly
+    ./install.sh --prefix=$SCRIPT_LOCATION >> $SCRIPT_LOCATION/log.txt 2>&1
+fi
+
 
 # CRLibM
 echo Installing CRLibM
@@ -36,9 +40,9 @@ mkdir -p crlibm && tar -xf crlibm.tar.gz -C crlibm --strip-components 1
 cd crlibm
 export CFLAGS=-fPIC $CFLAGS
 export LDFLAGS=-fPIC $LDFLAGS
-./configure --enable-sse2 --prefix=$SCRIPT_LOCATION >> $SCRIPT_LOCATION/log.txt
-make &>> log.txt
-make install >> log.txt
+./configure --enable-sse2 --prefix=$SCRIPT_LOCATION >> $SCRIPT_LOCATION/log.txt 2>&1
+make >> log.txt 2>&1
+make install >> log.txt 2>&1
 export LIBRARY_PATH=$SCRIPT_LOCATION/lib:$LIBRARY_PATH
 export C_INCLUDE_PATH=$SCRIPT_LOCATION/include:$C_INCLUDE_PATH
 export CPLUS_INCLUDE_PATH=$SCRIPT_LOCATION/include:$CPLUS_INCLUDE_PATH
@@ -47,17 +51,17 @@ export CPLUS_INCLUDE_PATH=$SCRIPT_LOCATION/include:$CPLUS_INCLUDE_PATH
 echo Installing GAOL
 cd $SOURCE_LOCATION
 rm -f gaol.tar.gz
-wget http://downloads.sourceforge.net/project/gaol/gaol/4.2.0/gaol-4.2.0.tar.gz -O gaol.tar.gz &>> $SCRIPT_LOCATION/log.txt
+wget http://downloads.sourceforge.net/project/gaol/gaol/4.2.0/gaol-4.2.0.tar.gz -O gaol.tar.gz >> $SCRIPT_LOCATION/log.txt 2>&1
 tar xf gaol.tar.gz
-patch -p0 < $SCRIPT_LOCATION/../documents/gaol-4.2.0.patch >> $SCRIPT_LOCATION/log.txt
+patch -p0 < $SCRIPT_LOCATION/../documents/gaol-4.2.0.patch >> $SCRIPT_LOCATION/log.txt 2>&1
 mv gaol-4.2.0 gaol
 cd gaol
 export CFLAGS=" -msse3 "; export CXXFLAGS=" -msse3 -std=c++11";
 ./configure  --with-mathlib=crlibm --enable-simd --enable-preserve-rounding=yes\
 	     --disable-debug --enable-optimize --disable-verbose-mode \
-	     --prefix=$SCRIPT_LOCATION &>> $SCRIPT_LOCATION/log.txt
-make &>> $SCRIPT_LOCATION/log.txt
-make install >> $SCRIPT_LOCATION/log.txt
+	     --prefix=$SCRIPT_LOCATION >> $SCRIPT_LOCATION/log.txt 2>&1
+make >> $SCRIPT_LOCATION/log.txt 2>&1
+make install >> $SCRIPT_LOCATION/log.txt 2>&1
 
 # Cleanup
 cd $SCRIPT_LOCATION
